+++ public/js/instruments/instrument-manager.js
@@ -812,15 +812,10 @@
          * @protected
          * @param {string} str - String to escape
          * @returns {string} Escaped string
+         * @see core/security.js - Delegates to central implementation
          */
         _escapeHtml(str) {
-            if (!str) return '';
-            return String(str)
-                .replace(/&/g, '&amp;')
-                .replace(/</g, '&lt;')
-                .replace(/>/g, '&gt;')
-                .replace(/"/g, '&quot;')
-                .replace(/'/g, '&#039;');
+            return global.SitesSecurity?.escapeHtml?.(str) ?? (str ? String(str).replace(/[&<>"']/g, m => ({'&':'&amp;','<':'&lt;','>':'&gt;','"':'&quot;',"'":'&#039;'})[m]) : '');
         }
 
         /**